Remove Cave(wo)man role and Sceptre of Might
Nobody in the community likes to play cavepeople. It's a bland melee
race that shouldn't exist alongside Bar and Sam.

The Sceptre was a bit too strong (MR and conflict?), and if I make it a
non-quest artifact it's even stronger (easier to wish for). I don't
really want to rebalance it, so it's gone for now.